requirements.txt
annotated-types==0.7.0
anyio==4.5.0
astroid==3.2.4
click==8.1.7
dill==0.3.8
fastapi==0.115.0
h11==0.14.0
idna==3.10
isort==5.13.2
mccabe==0.7.0
peewee==3.17.6
platformdirs==4.3.6
pydantic==2.9.2
pydantic_core==2.23.4
pylint==3.2.7
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.38.5
tomlkit==0.13.2
typing_extensions==4.12.2
uvicorn==0.30.6
